feat: Add organizational logo customization
This change introduces the ability to set and display a custom organizational logo throughout the application. The `logoUrl` state is added to the `AppContext` and persisted in local storage. The `Layout` and `Login` components are updated to display this logo, replacing the hardcoded default. An `updateLogo` function is added to `AppContext` and a new settings tab in `AdminDashboard` allows users to upload and set a new logo. This enhances branding and customization for different organizations.