feat: Refactor roles to support dynamic string types
This change modifies the `User` interface and `AuthContextType` to use string arrays for `roles` and `activeRole` instead of the `Role` enum. This allows for more flexibility in defining and managing roles dynamically.

Additionally, static lists for departments, positions, and roles have been added to `SystemLists` and populated in the `AppContext` to support a more configurable system. The `AdminDashboard` has also been updated to leverage these new lists for management.